refactor:make the four pictures into one picture
