--- frontend/src/app/dora/components/config-stepper/reports/average-cycle-time/average-cycle-time.component.html
@@ -1,3 +1,0 @@
-<div class="averageCycleTime-container">
-  <div id="averageCycleTime" style="width: 500px; height: 375px"></div>
-</div>
