--- frontend/src/app/dora/components/config-stepper/reports/average-cycle-time/average-cycle-time.component.scss
@@ -1,4 +1,0 @@
-.averageCycleTime-container {
-  width: 500px;
-  height: 375px;
-}
