--- frontend/src/app/dora/components/config-stepper/reports/block-reason-percentage/block-reason-percentage.component.html
@@ -1,3 +1,0 @@
-<div class="blockReasonPercentage-container">
-  <div id="blockReasonPercentage" style="width: 500px; height: 375px"></div>
-</div>
