--- frontend/src/app/dora/components/config-stepper/reports/block-reason-percentage/block-reason-percentage.component.scss
@@ -1,4 +1,0 @@
-.blockReasonPercentage-container {
-  width: 500px;
-  height: 375px;
-}
