+++ frontend/src/app/dora/components/config-stepper/reports/secondary-metrics/secondary-metrics.component.html
@@ -1,0 +1,4 @@
+<div class="secondary-metrics-container">
+  <app-divider text="Secondary metrics"></app-divider>
+  <div id="secondary-metrics" style="width: 1000px; height: 900px"></div>
+</div>
