+++ frontend/src/app/dora/components/config-stepper/reports/secondary-metrics/secondary-metrics.component.scss
@@ -1,0 +1,6 @@
+.secondary-metrics-container {
+  width: 1000px;
+  height: 900px;
+  margin: auto;
+  margin-bottom: 5%;
+}
