--- frontend/src/app/dora/components/config-stepper/reports/throughput/throughput.component.html
@@ -1,3 +1,0 @@
-<div class="throughput-container">
-  <div id="throughput" style="width: 500px; height: 375px"></div>
-</div>
