--- frontend/src/app/dora/components/config-stepper/reports/throughput/throughput.component.scss
@@ -1,4 +1,0 @@
-.throughput-container {
-  width: 500px;
-  height: 375px;
-}
