--- frontend/src/app/dora/components/config-stepper/reports/time-allocation/time-allocation.component.html
@@ -1,3 +1,0 @@
-<div class="time-allocation-container">
-  <div id="timeAllocation" style="width: 500px; height: 375px"></div>
-</div>
