--- frontend/src/app/dora/components/config-stepper/reports/time-allocation/time-allocation.component.scss
@@ -1,4 +1,0 @@
-.time-allocation-container {
-  width: 500px;
-  height: 375px;
-}
